FrogAi/FrogPilot FrogPilot-Staging 2024.11.26 01:15

+++ selfdrive/frogpilot/assets/theme_manager.py
@@ -63,6 +63,9 @@ def update_wheel_image(image, holiday_theme=None, random_event=True):
   else:
     wheel_location = os.path.join(THEME_SAVE_PATH, "steering_wheels")
 
+  if not os.path.exists(wheel_location):
+    return
+
   if os.path.exists(wheel_save_location):
     shutil.rmtree(wheel_save_location)
   os.makedirs(wheel_save_location, exist_ok=True)
